+++ packages/ui5-application-inquirer/test/unit/prompts/validators.test.ts
@@ -4,6 +4,7 @@ import { join } from 'path';
 import { initI18nUi5AppInquirer, t } from '../../../src/i18n';
 import { validateAppName, validateFioriAppProjectFolder } from '../../../src/prompts/validators';
 import { findRootsForPath } from '@sap-ux/project-access';
+import * as projectAccess from '@sap-ux/project-access';
 
 /**
  * Workaround to allow spyOn
@@ -16,7 +17,8 @@ jest.mock('@sap-ux/project-input-validator', () => {
 });
 
 jest.mock('@sap-ux/project-access', () => ({
-    findRootsForPath: jest.fn()
+    findRootsForPath: jest.fn(),
+    isPathForCapApp: jest.fn()
 }));
 
 describe('validators', () => {
@@ -56,7 +58,7 @@ describe('validators', () => {
         });
 
         test('should return true if no Fiori project is found in the target directory', async () => {
-            mockFindRootsForPath.mockResolvedValue(null);
+            mockFindRootsForPath.mockResolvedValueOnce(null);
             const result = await validateFioriAppProjectFolder('/path/to/dir');
             expect(result).toBe(true);
             expect(mockFindRootsForPath).toHaveBeenCalledWith('/path/to/dir');
@@ -65,10 +67,17 @@ describe('validators', () => {
         test('should return an error message if a Fiori project is found in the target directory', async () => {
             const appRootPath = '/path/to/fiori/project';
             const projectRootPath = 'test/path';
-            mockFindRootsForPath.mockResolvedValue({ appRoot: appRootPath, projectRoot: projectRootPath });
+            mockFindRootsForPath.mockResolvedValueOnce({ appRoot: appRootPath, projectRoot: projectRootPath });
             const result = await validateFioriAppProjectFolder('some/path');
             expect(result).toEqual(t('validators.folderContainsFioriApp', { path: appRootPath }));
             expect(mockFindRootsForPath).toHaveBeenCalledWith('some/path');
+        });
+
+        test('should return an error message if a CAP project is found in the target directory', async () => {
+            mockFindRootsForPath.mockResolvedValueOnce(null);
+            jest.spyOn(projectAccess, 'isPathForCapApp').mockResolvedValue(true);
+            const result = await validateFioriAppProjectFolder('any/path');
+            expect(result).toEqual(t('validators.folderContainsCapApp'));
         });
     });
 });